+++ CHANGELOG.md
@@ -6,7 +6,7 @@ CHANGELOG for Sulu
     * HOTFIX      #3793 [PreviewBundle]           Set context data (token, locale) before render can fail
     * ENHANCEMENT #3779 [ContentBundle]           Improved cache-invalidation for categories/tags in excerpt tab
     * ENHANCEMENT #3777 [ContentBundle]           Added tag/category reference-store
-    * ENHANCEMENT #3778 [ContactBundle]            Improved cache-invalidation for media/contact/account
+    * ENHANCEMENT #3778 [ContactBundle]           Improved cache-invalidation for media/contact/account
     * ENHANCEMENT #3778 [ContactBundle]           Added reference-store to contact/account
     * ENHANCEMENT #3778 [MediaBundle]             Added reference-store to media
     * FEATURE     #3028 [MediaBundle]             Added blur, grayscale, gamma, negative and sharpen transformation to media